LPS-56692 Consistency: Ant and Maven task names should start with a verb
