+++ modules/util/css-builder/samples/build.xml
@@ -10,7 +10,7 @@
 	<taskdef classpathref="css.builder.classpath" resource="com/liferay/css/builder/ant/taskdefs.properties" />
 
 	<target name="build-css">
-		<css-builder
+		<build-css
 			docrootDirName="css"
 			portalCommonDirName="../../../modules/frontend/frontend-common-css/tmp/META-INF/resources"
 		/>
