+++ modules/util/css-builder/src/main/resources/com/liferay/css/builder/ant/taskdefs.properties
@@ -1,1 +1,1 @@
-css-builder=com.liferay.css.builder.ant.CSSBuilderTask
+build-css=com.liferay.css.builder.ant.BuildCSSTask
